Added similarity visualizer
Added similarity visualizer for each song property. The similarity score is pretty simple, just original/current, with conditions to make sure it gives a useful value. This similarity score is used to draw a shade of green that corresponds to the % similarity. Also displays the percentage on mouseover
Fixed small issues with radio buttons